Add new e2e test for this remediation
- Refactor actual existing remediation test into own file
- Call new metal3remediation test from remediation_based_feature_test.go
